refactor(frontend): consolidate product hooks into useProductsWorkspace
